XAI-Multimodel/requirements.txt
flask
torch
torchvision
numpy
opencv-python
Pillow
matplotlib